Remove repository field from composer.json
The reprository field in the composer.json is used to set alternative
repositories for root dependencies. The current value is invalid, as
it violates the schema (https://getcomposer.org/doc/04-schema.md#repositories)
and is used for information that does not belong in this field. (It is
redundant, since the package is registered on packagist.org)